warn_failure_to_open_read_optional_path(): a new helper
Often we try to open a file for reading whose existence is optional,
and silently go on with our life when open() or fopen() fails, as if
nothing happened, because such a file is optional.  But open() or
fopen() may be failing not because the file is not there, but for
other reasons (like IO error or insufficient permission).

Add a helper for such codepaths to call after they fail to open/fopen
an optional file for reading, which would call warn_on_inaccessible()
unless the file in question is missing.

Signed-off-by: Junio C Hamano <gitster@pobox.com>